EXTREME graphics makeover
I fully abstracetd the graphics that we use, and introduced primitive
shapes. For a proof of product, I made the implementation for both LWJGL
and processing. It was actually hard at first to use processing, since
it has its own java executable and its own repeating loop which meant
that our loop was never called.